Split model code out from api-client
