+++ app/api-client.js
@@ -96,196 +96,3 @@ const api = new APIClient(config.api);
 api.auth = new JWTAuth();
 
 exports.api = api;
-
-
-const model_proxy = {
-  get: (model, property) => {
-    if (Reflect.has(model, property)) {
-      return Reflect.get(model, property);
-    }
-    if (model.data && property in model.data) {
-      return model.data[property];
-    }
-    return undefined;
-  },
-};
-
-
-class ModelSet extends Array {
-  constructor(Model, data = []) {
-    super(...data.map(obj => new Model(obj)));
-    this.model = Model;
-  }
-
-  exclude(other) {
-    let others = other;
-    if (other instanceof this.model) {
-      others = [other];
-    }
-
-    const pks = others.map(instance => instance[this.model.pk]);
-
-    return this.filter(instance => !pks.includes(instance[this.model.pk]));
-  }
-}
-
-exports.ModelSet = ModelSet;
-
-class Model {
-  constructor(data) {
-    this.data = data;
-    return new Proxy(this, model_proxy);
-  }
-
-  static get pk() {
-    return 'id';
-  }
-
-  static list() {
-    return api.get(this.endpoint)
-      .then(result => new ModelSet(this.prototype.constructor, result.results));
-  }
-
-  static get(id) {
-    return api.get(`${this.endpoint}/${id}`)
-      .then(data => new this.prototype.constructor(data));
-  }
-
-  static delete(id) {
-    return api.delete(`${this.endpoint}/${id}`);
-  }
-
-  create() {
-    return api.post(this.constructor.endpoint, this.data)
-      .then(data => new this.constructor(data));
-  }
-
-  replace() {
-    const pk_name = this.constructor.pk;
-    const pk = this.data[pk_name];
-
-    if (pk !== undefined) {
-      return api.put(`${this.constructor.endpoint}/${pk}`, this.data)
-        .then(data => new this.constructor(data));
-    }
-
-    return Promise.reject(new Error(`Missing ${pk_name} for PUT ${this.constructor.endpoint}`));
-  }
-
-  update() {
-    const pk_name = this.constructor.pk;
-    const pk = this.data[pk_name];
-
-    if (pk !== undefined) {
-      return api.patch(`${this.constructor.endpoint}/${pk}`, this.data)
-        .then(data => new this.constructor(data));
-    }
-
-    return Promise.reject(new Error(`Missing ${pk_name} for PATCH ${this.constructor.endpoint}`));
-  }
-}
-
-
-class AppS3Bucket extends Model {
-  static get endpoint() {
-    return 'apps3buckets';
-  }
-}
-
-exports.AppS3Bucket = AppS3Bucket;
-
-
-class UserS3Bucket extends Model {
-  static get endpoint() {
-    return 'users3buckets';
-  }
-}
-
-exports.UserS3Bucket = UserS3Bucket;
-
-
-class User extends Model {
-  static get endpoint() {
-    return 'users';
-  }
-
-  static get pk() {
-    return 'auth0_id';
-  }
-
-  get apps() {
-    return new ModelSet(App, this.data.userapps.map(ua => ua.app));
-  }
-
-  get buckets() {
-    return new ModelSet(Bucket, this.data.users3buckets.map(us => us.s3bucket));
-  }
-
-  get users3buckets() {
-    return new ModelSet(UserS3Bucket, this.data.users3buckets);
-  }
-}
-
-exports.User = User;
-
-
-class App extends Model {
-  static get endpoint() {
-    return 'apps';
-  }
-
-  get apps3buckets() {
-    return new ModelSet(AppS3Bucket, this.data.apps3buckets);
-  }
-
-  get buckets() {
-    return new ModelSet(Bucket, this.data.apps3buckets.map(as => as.s3bucket));
-  }
-
-  grant_bucket_access(bucket, access_level = 'readonly') {
-    if (!['readonly', 'readwrite'].includes(access_level)) {
-      throw new Error(`Invalid access_level "${access_level}"`);
-    }
-
-    let bucket_id = bucket;
-
-    if (typeof bucket === 'function' && bucket.prototype.constructor === Bucket) {
-      bucket_id = bucket.id;
-    }
-
-    return new AppS3Bucket({
-      app: this.id,
-      s3bucket: bucket_id,
-      access_level,
-    }).create();
-  }
-}
-
-exports.App = App;
-
-class Bucket extends Model {
-  static get endpoint() {
-    return 's3buckets';
-  }
-
-  get apps() {
-    return new ModelSet(App, this.data.apps3buckets.map(as => as.app));
-  }
-
-  get apps3buckets() {
-    return new ModelSet(AppS3Bucket, this.data.apps3buckets);
-  }
-
-  get users() {
-    return new ModelSet(User, this.data.users3buckets.map(us => us.user));
-  }
-
-  get users3buckets() {
-    return new ModelSet(UserS3Bucket, this.data.users3buckets);
-  }
-}
-
-exports.Bucket = Bucket;
-
-
-api.connect_bucket_to_app = apps3bucket => api.post('apps3buckets', body = apps3bucket);
