+++ test/test-app-edit.js
@@ -1,4 +1,5 @@
-const { App, Bucket, ModelSet, User } = require('../app/api-client');
+const { ModelSet } = require('../app/base-model');
+const { App, Bucket, User } = require('../app/models');
 const { assert } = require('chai');
 
 const { config, mock_api, url_for } = require('./conftest');
