+++ test/test-users.js
@@ -1,6 +1,7 @@
 "use strict";
 const { assert } = require('chai');
-const { User, ModelSet } = require('../app/api-client');
+const { ModelSet } = require('../app/base-model');
+const { User } = require('../app/models');
 const { config, mock_api } = require('./conftest');
